Update
when startup, minimize the terminal. Make so multiple accessories can be on at ounce. Also make it so files checked are also any folders, so scan folders. Because then, all the files can be kept inside the main "Mic Character APP" folder Name of folder should not matter, what should matter is where the main python file is located, if "micVis.py" is in a folder or section, add the other folders/files to it, this makes sure that files for the app are in the same place